requirements.txt
Django==4.2.4
djangorestframework==3.14.0
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
Pillow==9.2.0
Pillow==9.3.0
Pillow==9.1.1
Pillow==9.4.0
Pillow==10.0.0
torch==2.0.1
torch==2.0.1
torchvision==0.15.2
python==3.9.17